v4.2: Analytics page with Market Mood gauge (own breadth/trend/VIX inputs), breadth aggregates in run.json, clickable Track record outcome tiles + row drawer, compact universe/symbol_status JSON, route code-splitting; rescan 2026-09-16

+++ engine/types.py
@@ -396,7 +396,7 @@ class ScanResult:
     candidates: tuple[Candidate, ...]
     symbol_status: tuple[SymbolStatus, ...]
     universe: tuple[UniverseRow, ...]
-    counts: dict[str, int]
+    counts: dict[str, float]  # ints plus a few breadth medians
     warnings: tuple[str, ...]
     ban_list_available: bool
     scan_performed: bool
